Apply reviewed change
